Add suport for placeholders for grpc
